Move InitRepr into type_info.h
